* Skills and buff items should now expire properly.  * Refactored StatusEffectTools and the Player class' interface for applying effects. StatusEffectTools now only contains common methods that all types of status effects (i.e. buff skills, buff items, debuffs from monsters). The specific behaviors for the aforementioned status effects are now split into their own classes.  * Mob skill IDs are now represented internally as 2-bytes (short) instead of 4-bytes (int). In order to simplify compiling of Skill.wz in KvjCompiler, mob skill IDs are still 4-bytes on the KVJ files (and are just downcasted).  * Fixed unable to warp bug following a skill cast.  * Fixed reactor and player drops being dropped from the wrong place if a mob with the same entity ID was visible on the map.  * Added more definitions to PlayerJob.  * Made some progress on ranged attacks (still nothing on magic attacks yet).  * Rechargeable ammo (stars, bullets) are now a special case in the inventory manipulation functions in InventoryTools. Adding rechargeable items to the inventory will not change the quantity of non-full items that share the same item ID, and removing rechargeable items will not clear the slot if the new quantity equals 0.  * Quest requirement watching in Player should perform a bit better now.

+++ trunk/src/argonms/loading/skill/KvjSkillDataLoader.java
@@ -60,7 +60,7 @@ public class KvjSkillDataLoader extends SkillDataLoader {
 		}
 	}
 
-	protected void loadMobSkill(int skillid) {
+	protected void loadMobSkill(short skillid) {
 		try {
 			File f = new File(new StringBuilder(dataPath).append("Skill.wz").append(File.separator).append("MobSkill.img.kvj").toString());
 			if (f.exists())
@@ -96,7 +96,7 @@ public class KvjSkillDataLoader extends SkillDataLoader {
 	//TODO: Actually do real work to see if the skill exists in the file so we
 	//can name this method exists() instead of loadable?
 	public boolean canLoadPlayerSkill(int skillid) {
-		if (skillStats.containsKey(skillid))
+		if (skillStats.containsKey(Integer.valueOf(skillid)))
 			return true;
 		String id = String.format("%07d", skillid);
 		File f = new File(new StringBuilder(dataPath).append("Skill.wz").append(File.separator).append(id.substring(0, 3)).append(".img.kvj").toString());
@@ -105,8 +105,8 @@ public class KvjSkillDataLoader extends SkillDataLoader {
 
 	//TODO: Actually do real work to see if the skill exists in the file so we
 	//can name this method exists() instead of loadable?
-	public boolean canLoadMobSkill(int skillid) {
-		if (mobSkillStats.containsKey(skillid))
+	public boolean canLoadMobSkill(short skillid) {
+		if (mobSkillStats.containsKey(Short.valueOf(skillid)))
 			return true;
 		File f = new File(new StringBuilder(dataPath).append("Skill.wz").append(File.separator).append("MobSkill.img.kvj").toString());
 		return f.exists();
@@ -259,13 +259,13 @@ public class KvjSkillDataLoader extends SkillDataLoader {
 	private void doMobWork(LittleEndianReader reader) {
 		MobSkillStats stats = null;
 		byte level;
-		int skillid = -1;
+		short skillid = -1;
 		for (byte now = reader.readByte(); now != -1; now = reader.readByte()) {
 			switch (now) {
 				case NEXT_SKILL:
 					stats = new MobSkillStats();
-					skillid = reader.readInt();
-					mobSkillStats.put(Integer.valueOf(skillid), stats);
+					skillid = (short) reader.readInt();
+					mobSkillStats.put(Short.valueOf(skillid), stats);
 					break;
 				case ELEM_ATTR:
 					stats.setElemAttr(reader.readNullTerminatedString());
@@ -287,7 +287,7 @@ public class KvjSkillDataLoader extends SkillDataLoader {
 		}
 	}
 
-	private MobSkillEffectsData processMobEffect(int skillid, byte level, LittleEndianReader reader) {
+	private MobSkillEffectsData processMobEffect(short skillid, byte level, LittleEndianReader reader) {
 		MobSkillEffectsData effect = new MobSkillEffectsData(skillid, level);
 		loop:
 		for (byte now = reader.readByte(); now != -1; now = reader.readByte()) {
